requirements.txt
langchain
langgraph
faiss-cpu
python-dotenv
requests
pydantic-settings
langchain-ollama
pydantic_ai
